requirements.txt
Flask~=2.2.3
Flask-WTF~=1.1.1
Jinja2~=3.1.2
MarkupSafe~=2.1.2
WTForms~=3.0.1
Werkzeug~=2.2.3
click~=8.1.3
dnspython~=2.3.0
email-validator~=1.3.1
idna~=3.4
itsdangerous~=2.1.2
mysql-connector-python~=8.0.32
pip~=22.3.1
protobuf~=3.20.3
setuptools~=65.5.1
wheel~=0.38.4
pyotp~=2.8.0
django-qrcode~=0.3